+++ examples/ex_static_lib/libmsg.h
@@ -1,0 +1,5 @@
+#pragma once
+char *get_msg_one(void);
+char *get_msg_two(void);
+char *get_msg_three(void);
+char *get_msg_four(void);