+++ examples/ex_static_lib/msg_one.c
@@ -1,0 +1,6 @@
+#include "libmsg.h"
+
+char *get_msg_one(void)
+{
+    return "one";
+}