+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 1.5.2
 * Add ELF x86_64 .o support.
+* Add static archive support.
 
 # 1.5.1
 * Use dedicated symbolic expression attributes.